+++ CHANGELOG.md
@@ -18,3 +18,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `Game` class.
 - `Engine` class.
 - Take inputs interchangebly from players.
+- defining row, column and diagonal? of last move in `Engine` class.